# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Saina Gallery RESTful API",
        "description": "SainaGallery API documentation"
    },
...
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/superadmin/changeblock": {
            "get": {
                "tags": [
                    "superAdmin"
                ],
                "summary": "Change User Block Condition",
                "description": "Block Condition",
                "operationId": "changeuserblockcondition",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "condition",
                        "in": "query",
                        "type": "boolean"
                    },
                    {
                        "name": "id",
                        "type": "string",
                        "in": "query",
                        "description": "User ID"
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/api/superadmin/province": {
            "post": {
                "tags": [
                    "superAdmin"
                ],
                "summary": "Add New Province",
                "description": "Add Province",
                "operationId": "newprovince",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "Add Province",
                        "description": "Add new Province",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/province"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "superAdmin"
                ],
                "summary": "Edit a Province",
                "description": "Edit Province",
                "operationId": "editcountry",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "edit Province",
                        "description": "Add new Province",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/province"
                        }
                    },
                    {
                        "in": "query",
                        "name": "id",
                        "type": "string"
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "superAdmin"
                ],
                "summary": "Delete one Province",
                "description": "Delete one Province",
                "operationId": "deleteprovince",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "id",
                        "description": "Delete one Province by id",
                        "type": "string"
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/api/superadmin/city": {
            "post": {
                "tags": [
                    "superAdmin"
                ],
                "summary": "Add New City",
                "description": "Add City",
                "operationId": "newcity",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "Add City",
                        "description": "Add new City",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/city"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "superAdmin"
                ],
                "summary": "Edit a City",
                "description": "Edit City",
                "operationId": "editcity",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "Add City",
                        "description": "Add new City",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/city"
                        }
                    },
                    {
                        "in": "query",
                        "name": "id",
                        "type": "string"
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "superAdmin"
                ],
                "summary": "Delete one City",
                "description": "Delete one City",
                "operationId": "deletecity",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "id",
                        "description": "Delete one City by id",
                        "type": "string"
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/api/superadmin/category": {
            "post": {
                "tags": [
                    "superAdmin"
                ],
                "summary": "Add New Category",
                "description": "Add Category",
                "operationId": "newcategory",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "type": "string",
                        "in": "formData"
                    },
                    {
                        "name": "categoryImage",
                        "type": "file",
                        "in": "formData"
                    },
                    {
                        "name": "description",
                        "type": "string",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "superAdmin"
                ],
                "summary": "Edit a Category",
                "description": "Edit Category",
                "operationId": "editcategory",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "type": "string",
                        "in": "formData"
                    },
                    {
                        "name": "categoryImage",
                        "type": "file",
                        "in": "formData"
                    },
                    {
                        "name": "id",
                        "type": "string",
                        "in": "query"
                    },
                    {
                        "name": "description",
                        "type": "string",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "superAdmin"
                ],
                "summary": "Delete a Category",
                "description": "Delete a Category",
                "operationId": "deletecategory",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "id",
                        "description": "Delete one Category by id",
                        "type": "string"
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
...
    },
    "securityDefinitions": {
        "Authorization": {
            "type": "apiKey",
            "in": "header",
            "name": "Authorization"
        }
    },
    "definitions": {
...
        "province": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "کرمانشاه"
                }
            },
            "xml": {
                "name": "province"
            }
        },
        "city": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "سرپل ذهاب"
                },
                "province": {
                    "type": "string",
                    "example": "province id"
                }
            },
            "xml": {
                "name": "city"
            }
        },
...
    }
}